+++ priv/ts/css-update-module.ts
@@ -1,4 +1,7 @@
-import { removeStyle as __volt_removeStyle, updateStyle as __volt_updateStyle } from '/@volt/client.js'
+import {
+  removeStyle as __volt_removeStyle,
+  updateStyle as __volt_updateStyle
+} from '/@volt/client.js'
 
 const __volt_id = $id
 const __volt_css = $css